% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3bin.R
\name{bin_from_lonlat}
\alias{bin_from_lonlat}
\title{Generate bin number from longitude latitude.}
\usage{
bin_from_lonlat(lon, lat, NUMROWS)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{NUMROWS}{number of rows in the grid}
}
\value{
integer vector of bin number
}
\description{
Bin number from longitude and latitude for a given grid with NUMROWS unique latitudes.
}
\details{
This function previously lived in sosoc/croc where it is called lonlat2bin
}
\examples{
bin_from_lonlat(147, -42, 1024)
bin_from_lonlat(c(0, 0, 0), c(-90, 0, 90), 1024)
}
