% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3bin.R
\name{crop_bins}
\alias{crop_bins}
\title{Crop L3 init object with an extent}
\usage{
crop_bins(x, extent)
}
\arguments{
\item{x}{L3bin object}

\item{extent}{vector of 'c(xmin, xmax, ymin, ymax)'}
}
\value{
integer vector of bins
}
\description{
Crop L3 list, returns bins that fall within the extent.
}
\details{
duplicated from sosoc/croc where it is called crop_init
}
\examples{
init <- L3bin(24)
crop_bins(init, c(100, 110, -50, -45))
}
