% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3bin.R
\name{extent_from_bin}
\alias{extent_from_bin}
\title{Calculate bin boundaries from bin number}
\usage{
extent_from_bin(bin, NUMROWS)
}
\arguments{
\item{bin}{bin number}

\item{NUMROWS}{relevant number of L3 bin rows}
}
\description{
Calculate bin boundaries from bin number
}
\details{
Originally from sosoc/croc where it is called bin2bounds
}
