% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3bin.R
\name{L3bin}
\alias{L3bin}
\title{Set up the basic values for the bin scheme for given number of rows.}
\usage{
L3bin(NUMROWS = 2160)
}
\arguments{
\item{NUMROWS}{relevant number of L3 bin rows}
}
\description{
Originally from sosoc/croc where it is called initbin
}
\examples{
L3bin(1024)
}
\references{
https://oceancolor.gsfc.nasa.gov/docs/format/l3bins/
}
