% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3bin.R
\name{lonlat_from_bin}
\alias{lonlat_from_bin}
\title{Longitude and latitude from bin number.}
\usage{
lonlat_from_bin(bins, nrows)
}
\arguments{
\item{bins}{bin number}

\item{nrows}{number of rows in this grid}
}
\description{
Generate longitude and latitude coordinates from bin number.
}
\details{
Originally from sosoc/croc where it is called bin2lonlat
}
