% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3bin.R
\name{row_from_lat}
\alias{row_from_lat}
\title{Latitude to row}
\usage{
row_from_lat(lat, NUMROWS)
}
\arguments{
\item{lat}{latitude}

\item{NUMROWS}{number of rows in the grid}
}
\description{
Row is 1-based, and starts at the southern-most row.
}
\details{
This function previously lived in sosoc/croc where it is called .lat2row.
}
\examples{
row_from_lat(-42, 1024)
row_from_lat(c(-90, 0, 90), 1024)
}
